requirements.txt
asgiref==3.7.2
certifi==2024.6.2
charset-normalizer==3.3.2
Django==5.0.2
django-extensions==3.2.3
django-filter==23.5
django-unfold==0.28.1
djangorestframework==3.14.0
djangorestframework-jsonapi==6.1.0
djangorestframework-simplejwt==5.3.1
idna==3.7
inflection==0.5.1
mysql==0.0.3
mysql-connector-python==8.4.0
mysqlclient==2.2.4
pillow==10.2.0
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
rest-framework-simplejwt==0.0.2
sqlparse==0.4.4
tzdata==2024.1
urllib3==2.2.2
